% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTM_counts.R
\name{DTM_counts}
\alias{DTM_counts}
\title{Convert leaf counts to node or branch counts for DTM variable.}
\usage{
DTM_counts(Y, tree, branches = FALSE, prop = FALSE)
}
\arguments{
\item{Y}{Numeric matrix of DTM observations}

\item{tree}{An object of class \code{ape} specificying the DTM structure of Y.}

\item{branches}{A logical (default = \code{FALSE}) indicating if counts should be in branch order.}

\item{prop}{A logical (default = \code{FALSE}) indicating if counts should be converted to proportions.}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Convert leaf counts to node or branch counts for DTM variable.
}
