% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTM_likelihood.R
\name{DTM_likelihood}
\alias{DTM_likelihood}
\title{Calculate likelihood for DTM random variable}
\usage{
DTM_likelihood(y, gamma, tree, log = TRUE, simplify = TRUE)
}
\arguments{
\item{y}{Numeric matrix of DTM count values.}

\item{gamma}{Numeric matrix of Dirichlet parameters for DTM variable}

\item{tree}{An object of class \code{ape} specificying the DTM structure of y.}

\item{log}{Logical indicating if log-likelihood should be returned (Default = \code{TRUE})}

\item{simplify}{Logical. TRUE (default): return overall likelihood; FALSE: return observation-wise likelihoods}
}
\value{
A numeric vector
}
\description{
Calculate likelihood for DTM random variable
}
