% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_DTM.R
\name{predict_DTM}
\alias{predict_DTM}
\title{Produces expected counts for each node for specfied DTM variables.}
\usage{
predict_DTM(gamma, tree, N_i = NULL)
}
\arguments{
\item{gamma}{Numeric matrix of Dirichlet parameter values.}

\item{tree}{An object of class \code{ape} specificying the DTM structure of Y.}

\item{N_i}{Optional numeric vector of sample sizes. If \code{NULL} (default), will return proportions.}
}
\value{
A numeric \code{matrix}
}
\description{
Produces expected counts for each node for specfied DTM variables.
}
