% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_DTM.R
\name{sample_DTM}
\alias{sample_DTM}
\title{Sample from a DTM distribution}
\usage{
sample_DTM(gamma, N, tree)
}
\arguments{
\item{gamma}{Numeric matrix of dirichlet parameters for each branch.}

\item{N}{Numeric vector of sample sizes for each observation.}

\item{tree}{An object of class \code{ape} specificying the DTM structure of Y.}
}
\value{
A numeric \code{matrix} object with samples'
}
\description{
Sample from a DTM distribution
}
