% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_DTM.R
\name{simulate_DTM}
\alias{simulate_DTM}
\title{Simulate a DTM for use with treesharp_stan function.}
\usage{
simulate_DTM(
  subject_sim = 100,
  subject_N = NULL,
  tree = NULL,
  num_leaf = 5,
  covariates_sim = 50,
  which_coninuous = NULL,
  which_binary = NULL,
  binary_p = NULL,
  zeta = NULL,
  rho = 0.3,
  Sigma = NULL,
  num_cov = 5,
  beta_min = 0.8,
  beta_max = 1.2,
  seed = 1234
)
}
\arguments{
\item{subject_sim}{number of subjects/groups/observations to simulate}

\item{subject_N}{vector of length(subject_sim with group sizes. NULL (default) will randomly sample from Unif(7000, 10000)}

\item{tree}{object of type \code{ape}}

\item{num_leaf}{if tree is NULL, number of leaves on generated tree}

\item{covariates_sim}{number of covariates to generate (i.e. ncol(X))}

\item{which_binary}{indexes of binary covariates}

\item{binary_p}{success probability for binary draws}

\item{zeta}{matrix indicating nonzero variables for each branch}

\item{rho}{correlation for continuous covariates at lag 1. Either rho or Sigma must be specified.}

\item{Sigma}{correlation matrix for continuous covariates. Either rho or Sigma must be specified.}

\item{num_cov}{number of nonzero regression coefficients per branch}

\item{beta_min}{minimum absolute value for regression coefficients}

\item{beta_max}{maximum absolute value for regression coefficients}

\item{seed}{random seed}

\item{which_continuous}{indexes of continuous covariates}
}
\value{
A list with, X, Y, alpha, beta, zeta, tree
}
\description{
Simulate a DTM for use with treesharp_stan function.
}
