% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesharp_stan_tree_inputs.R
\name{treesharp_stan_tree_inputs}
\alias{treesharp_stan_tree_inputs}
\title{Support function for treesharp_stan}
\usage{
treesharp_stan_tree_inputs(tree)
}
\arguments{
\item{tree}{An object of class \code{ape} specificying the DTM structure of Y.}
}
\value{
A \code{list} with 6 element
}
\description{
Support function for treesharp_stan
}
